+++ CHANGELOG.md
@@ -3,9 +3,7 @@
         New features will be added to this release over time, the current changelog is not final.
 
 ### To-Do
- - Delete old tensorrt models
  - Fix FFMpeg Colors
- - Fix Upscaling Preview
 
 ### Added
  - Batch input support.
@@ -15,6 +13,7 @@
  - Incorrect colors with some videos.
  - Incorrect RIFE 4K output
  - Upscaling TensorRT on Windows.
+ - Upscaling Preview.
 # RVE 2.3.2
 ### Added
  - Fallback to float32 if float16 does not work with custom arch.